+++ src/core/auth/auth-service.models.ts
@@ -1,0 +1,7 @@
+import { FirebaseAuthTypes } from '@react-native-firebase/auth'
+
+export interface AuthService {
+  init: () => void;
+  googleSignIn: () => Promise<false | FirebaseAuthTypes.UserCredential>;
+  getCurrentUser: () => FirebaseAuthTypes.User | null;
+}