+++ src/dictionarly/components/dictionarly-play/dictionarly-play.styled.ts
@@ -6,9 +6,14 @@ import { COLOR } from '@core/colors/colors.constants'
 import { TEXT_SIZE } from '@core/text/text.constants'
 import { getThemeProp } from '@core/styled/theme'
 import { ThemeModel } from '@core/styled/models'
+import { isPlatform } from '@core/is-platform/is-platform'
 
-export const DictionarlyContainer = styled.View`
-  justify-content: center;
+export const DictionarlyContainer = styled.ScrollView.attrs({
+  contentContainerStyle: {
+    justifyContent: 'center',
+    height: '100%',
+  },
+})`
   flex: 1;
 `
 
@@ -56,3 +61,9 @@ export const DictionarlySendButtonIcon = styled(MaterialCommunityIcons).attrs({
   color: COLOR.DODGER_BLUE,
   size: TEXT_SIZE.XXXL,
 })``
+
+export const DictionarlyKeyboardAvoidingView = styled.KeyboardAvoidingView.attrs({
+  behavior: isPlatform('ios') ? 'padding' : 'height',
+})`
+  flex: 1;
+`