+++ src/more/more.navigation.tsx
@@ -7,6 +7,7 @@ import { Playground } from '../playground/playground'
 import { HelpNavigation } from '../help/help.navigation'
 import { Author } from './components/author/author'
 import { Developer } from '../developer/developer'
+import { User } from '../user/user'
 
 const Stack = createNativeStackNavigator()
 
@@ -19,6 +20,7 @@ export const MoreNavigation = () => {
       <Stack.Screen name={SCREEN.MORE_HELP} component={HelpNavigation} />
       <Stack.Screen name={SCREEN.MORE_AUTHOR} component={Author} />
       <Stack.Screen name={SCREEN.DEVELOPER} component={Developer} />
+      <Stack.Screen name={SCREEN.MORE_USER} component={User} />
     </Stack.Navigator>
   )
 }